requirements.txt
numpy
pandas
scikit-learn
torch
transformers
pytorch-lightning
neo4j
networkx
nltk
sentencepiece
spacy
indic-nlp-library
transformers